requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4  # https://github.com/spyder-ide/spyder/issues/22187
rdkit-pypi==2022.9.5
scipy==1.13.1
plotly==5.24.1
matplotlib==3.9.2
pubchempy
JPype1==1.5.0
